% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_mrf_logC_cpp}
\alias{calc_mrf_logC_cpp}
\title{Compute log normalizing constant for the edge-sharing MRF prior}
\usage{
calc_mrf_logC_cpp(Theta, nu)
}
\arguments{
\item{Theta}{A K x K symmetric matrix.}

\item{nu}{Numeric vector of length 1 or more.}
}
\value{
Numeric vector of \eqn{\log C(\Theta,\nu)} for each entry of \code{nu}.
}
\description{
Computes \eqn{\log C(\Theta, \nu)} where
\deqn{C(\Theta,\nu)=\sum_{g\in\{0,1\}^K} \exp\{\nu \mathbf{1}^\top g + g^\top \Theta g\}.}
}
