% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.multiggm_platform_fit}
\alias{fitted.multiggm_platform_fit}
\title{Extract posterior mean partial correlation matrices from a multi-platform fit}
\usage{
\method{fitted}{multiggm_platform_fit}(object, platform = NULL, ...)
}
\arguments{
\item{object}{A \code{multiggm_platform_fit} object.}

\item{platform}{Integer; which platform. If NULL, returns all.}

\item{...}{Ignored.}
}
\value{
A list of K posterior mean partial correlation matrices, or a list
of S such lists.
}
\description{
Extract posterior mean partial correlation matrices from a multi-platform fit
}
