% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor.R
\name{posterior_ci}
\alias{posterior_ci}
\title{Posterior credible intervals for partial correlations}
\usage{
posterior_ci(pcor, probs = c(0.025, 0.975))
}
\arguments{
\item{pcor}{array \link{p,p,K,nsave}}

\item{probs}{numeric length-2, e.g. c(0.025, 0.975)}
}
\value{
list with lower/median/upper arrays \link{p,p,K}
}
\description{
Posterior credible intervals for partial correlations
}
