% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor.R
\name{confusion_at_threshold}
\alias{confusion_at_threshold}
\title{Confusion metrics at a threshold}
\usage{
confusion_at_threshold(score_mat, truth_mat, thr, upper_only = TRUE)
}
\description{
Confusion metrics at a threshold
}
