% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor.R
\name{diff_prob_pcor}
\alias{diff_prob_pcor}
\title{Posterior probability of differential partial correlation (K=2)}
\usage{
diff_prob_pcor(pcor, delta = 0)
}
\arguments{
\item{pcor}{array \link{p,p,2,nsave}}

\item{delta}{practical difference threshold (0 = any difference)}
}
\value{
matrix \link{p,p} of posterior probabilities
}
\description{
Posterior probability of differential partial correlation (K=2)
}
