% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiggm_mcmc.R
\name{.multiggm_mcmc_single}
\alias{.multiggm_mcmc_single}
\title{Internal single-chain engine}
\usage{
.multiggm_mcmc_single(
  S_list,
  n_vec,
  burnin,
  nsave,
  thin,
  chain_id = 1L,
  Theta_init = NULL,
  nu_init = NULL,
  C_init = NULL,
  disp = FALSE,
  hyper = NULL,
  seed = NULL,
  ...
)
}
\description{
Implements the MATLAB routine \code{MCMC_multiple_graphs.m} using the Wang–Li
G-Wishart kernels (ported to RcppArmadillo).
}
\keyword{internal}
