% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor.R
\name{pip_diff_edge}
\alias{pip_diff_edge}
\title{Posterior inclusion probabilities (PIP) per group differential edges}
\usage{
pip_diff_edge(fit, chain = 1L)
}
\value{
array \link{p,p,K}
}
\description{
Posterior inclusion probabilities (PIP) per group differential edges
}
