% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor.R
\name{pip_edges}
\alias{pip_edges}
\title{Posterior inclusion probabilities (PIP) per group}
\usage{
pip_edges(fit, chain = 1L)
}
\arguments{
\item{fit}{a multi_ggm_fit_list object}
}
\value{
array \link{p,p,K}
}
\description{
Posterior inclusion probabilities (PIP) per group
}
